+++ sdk/src/client/middleware.rs
@@ -13,6 +13,18 @@ pub(crate) struct AdmissionPolicyMiddleware {
 }
 
 impl AdmitPolicy for AdmissionPolicyMiddleware {
+    fn admit_relay<'a>(
+        &'a self,
+        relay_url: &'a RelayUrl,
+    ) -> BoxedFuture<'a, Result<AdmitStatus, PolicyError>> {
+        Box::pin(async move {
+            match &self.external_policy {
+                Some(policy) => policy.admit_relay(relay_url).await,
+                None => Ok(AdmitStatus::Success),
+            }
+        })
+    }
+
     fn admit_connection<'a>(
         &'a self,
         relay_url: &'a RelayUrl,