+++ sdk/CHANGELOG.md
@@ -91,6 +91,7 @@
 - Add `GossipAllowedRelays` to `GossipOptions` to filter relays during selection (https://github.com/rust-nostr/nostr/pull/1128)
 - Add `AdmitPolicy::admit_auth` to control relay authentication (https://github.com/rust-nostr/nostr/pull/1218)
 - Add gossip background refresher (https://github.com/rust-nostr/nostr/pull/1260)
+- Add `AdmitPolicy::admit_relay` (https://github.com/rust-nostr/nostr/pull/1339)
 
 ## v0.44.1 - 2025/11/09